+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/living/types/ChickenShop.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.bukkit.Color;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
-import org.bukkit.entity.Pig;
+import org.bukkit.entity.Chicken;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.ItemStack;
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -26,15 +26,9 @@ import com.nisovin.shopkeepers.util.data.property.Property;
 import com.nisovin.shopkeepers.util.data.property.value.PropertyValue;
 import com.nisovin.shopkeepers.util.data.serialization.InvalidDataException;
 import com.nisovin.shopkeepers.util.data.serialization.bukkit.NamespacedKeySerializers;
-import com.nisovin.shopkeepers.util.data.serialization.java.BooleanSerializers;
 import com.nisovin.shopkeepers.util.inventory.ItemUtils;
 
-public class PigShop extends BabyableShop<Pig> {
-
-	public static final Property<Boolean> SADDLE = new BasicProperty<Boolean>()
-			.dataKeyAccessor("saddle", BooleanSerializers.LENIENT)
-			.defaultValue(false)
-			.build();
+public class ChickenShop extends BabyableShop<Chicken> {
 
 	// TODO Replace with the actual type once we only support MC 1.21.5+
 	public static final Property<NamespacedKey> VARIANT = new BasicProperty<NamespacedKey>()
@@ -42,17 +36,13 @@ public class PigShop extends BabyableShop<Pig> {
 			.defaultValue(NamespacedKey.minecraft("temperate"))
 			.build();
 
-	private final PropertyValue<Boolean> saddleProperty = new PropertyValue<>(SADDLE)
-			.onValueChanged(Unsafe.initialized(this)::applySaddle)
-			.build(properties);
-
 	private final PropertyValue<NamespacedKey> variantProperty = new PropertyValue<>(VARIANT)
 			.onValueChanged(Unsafe.initialized(this)::applyVariant)
 			.build(properties);
 
-	public PigShop(
+	public ChickenShop(
 			LivingShops livingShops,
-			SKLivingShopObjectType<PigShop> livingObjectType,
+			SKLivingShopObjectType<ChickenShop> livingObjectType,
 			AbstractShopkeeper shopkeeper,
 			@Nullable ShopCreationData creationData
 	) {
@@ -62,79 +52,26 @@ public class PigShop extends BabyableShop<Pig> {
 	@Override
 	public void load(ShopObjectData shopObjectData) throws InvalidDataException {
 		super.load(shopObjectData);
-		saddleProperty.load(shopObjectData);
 		variantProperty.load(shopObjectData);
 	}
 
 	@Override
 	public void save(ShopObjectData shopObjectData, boolean saveAll) {
 		super.save(shopObjectData, saveAll);
-		saddleProperty.save(shopObjectData);
 		variantProperty.save(shopObjectData);
 	}
 
 	@Override
 	protected void onSpawn() {
 		super.onSpawn();
-		this.applySaddle();
 		this.applyVariant();
 	}
 
 	@Override
 	public List<Button> createEditorButtons() {
 		List<Button> editorButtons = super.createEditorButtons();
-		editorButtons.add(this.getSaddleEditorButton());
 		editorButtons.add(this.getVariantEditorButton());
 		return editorButtons;
-	}
-
-	// SADDLE
-
-	public boolean hasSaddle() {
-		return saddleProperty.getValue();
-	}
-
-	public void setSaddle(boolean saddle) {
-		saddleProperty.setValue(saddle);
-	}
-
-	public void cycleSaddle() {
-		this.setSaddle(!this.hasSaddle());
-	}
-
-	private void applySaddle() {
-		Pig entity = this.getEntity();
-		if (entity == null) return; // Not spawned
-
-		entity.setSaddle(this.hasSaddle());
-	}
-
-	private ItemStack getSaddleEditorItem() {
-		ItemStack iconItem = new ItemStack(Material.SADDLE);
-		ItemUtils.setDisplayNameAndLore(
-				iconItem,
-				Messages.buttonPigSaddle,
-				Messages.buttonPigSaddleLore
-		);
-		return iconItem;
-	}
-
-	private Button getSaddleEditorButton() {
-		return new ShopkeeperActionButton() {
-			@Override
-			public @Nullable ItemStack getIcon(EditorSession editorSession) {
-				return getSaddleEditorItem();
-			}
-
-			@Override
-			protected boolean runAction(
-					EditorSession editorSession,
-					InventoryClickEvent clickEvent
-			) {
-				cycleSaddle();
-				return true;
-			}
-		};
 	}
 
 	// VARIANT
@@ -148,33 +85,33 @@ public class PigShop extends BabyableShop<Pig> {
 	}
 
 	public void cycleVariant(boolean backwards) {
-		this.setVariant(NMSManager.getProvider().cyclePigVariant(this.getVariant(), backwards));
+		this.setVariant(NMSManager.getProvider().cycleChickenVariant(this.getVariant(), backwards));
 	}
 
 	private void applyVariant() {
-		Pig entity = this.getEntity();
+		Chicken entity = this.getEntity();
 		if (entity == null) return; // Not spawned
 
-		NMSManager.getProvider().setPigVariant(entity, this.getVariant());
+		NMSManager.getProvider().setChickenVariant(entity, this.getVariant());
 	}
 
 	private ItemStack getVariantEditorItem() {
 		ItemStack iconItem = new ItemStack(Material.LEATHER_CHESTPLATE);
 		switch (this.getVariant().getKey()) {
 		case "warm":
-			ItemUtils.setLeatherColor(iconItem, Color.fromRGB(203, 114, 56));
+			ItemUtils.setLeatherColor(iconItem, Color.fromRGB(230, 178, 100));
 			break;
 		case "cold":
-			ItemUtils.setLeatherColor(iconItem, Color.fromRGB(226, 201, 148));
+			ItemUtils.setLeatherColor(iconItem, Color.fromRGB(116, 106, 125));
 			break;
 		case "temperate":
 		default:
-			ItemUtils.setLeatherColor(iconItem, Color.fromRGB(252, 183, 179));
+			ItemUtils.setLeatherColor(iconItem, Color.WHITE);
 			break;
 		}
 		ItemUtils.setDisplayNameAndLore(iconItem,
-				Messages.buttonPigVariant,
-				Messages.buttonPigVariantLore
+				Messages.buttonChickenVariant,
+				Messages.buttonChickenVariantLore
 		);
 		return iconItem;
 	}