+++ modules/v1_21_R5/src/main/java/com/nisovin/shopkeepers/compat/v1_21_R5/NMSHandler.java
@@ -1,8 +1,11 @@
 package com.nisovin.shopkeepers.compat.v1_21_R5;
 
 import java.lang.reflect.Field;
+import java.lang.reflect.Method;
 
 import org.bukkit.ExplosionResult;
+import org.bukkit.NamespacedKey;
+import org.bukkit.Registry;
 import org.bukkit.craftbukkit.v1_21_R4.entity.CraftAbstractVillager;
 import org.bukkit.craftbukkit.v1_21_R4.entity.CraftEntity;
 import org.bukkit.craftbukkit.v1_21_R4.entity.CraftLivingEntity;
@@ -13,9 +16,12 @@ import org.bukkit.craftbukkit.v1_21_R4.inventory.CraftItemStack;
 import org.bukkit.craftbukkit.v1_21_R4.inventory.CraftMerchant;
 import org.bukkit.craftbukkit.v1_21_R4.util.CraftMagicNumbers;
 import org.bukkit.entity.AbstractVillager;
+import org.bukkit.entity.Chicken;
+import org.bukkit.entity.Cow;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Mob;
+import org.bukkit.entity.Pig;
 import org.bukkit.entity.Player;
 import org.bukkit.entity.Salmon;
 import org.bukkit.entity.Villager;
@@ -30,13 +36,14 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import com.nisovin.shopkeepers.api.internal.util.Unsafe;
 import com.nisovin.shopkeepers.compat.api.NMSCallProvider;
 import com.nisovin.shopkeepers.shopobjects.living.LivingEntityAI;
+import com.nisovin.shopkeepers.util.bukkit.RegistryUtils;
 import com.nisovin.shopkeepers.util.inventory.ItemUtils;
 import com.nisovin.shopkeepers.util.java.EnumUtils;
 import com.nisovin.shopkeepers.util.java.Validate;
 import com.nisovin.shopkeepers.util.logging.Log;
 
-import net.minecraft.core.component.DataComponentMap;
 import net.minecraft.core.component.DataComponentExactPredicate;
+import net.minecraft.core.component.DataComponentMap;
 import net.minecraft.core.component.PatchedDataComponentMap;
 import net.minecraft.nbt.Tag;
 import net.minecraft.server.MinecraftServer;
@@ -48,10 +55,17 @@ import net.minecraft.world.item.trading.MerchantOffers;
 public final class NMSHandler implements NMSCallProvider {
 
 	private final Field craftItemStackHandleField;
+	private final Method cowSetVariantMethod;
 
 	public NMSHandler() throws Exception {
 		craftItemStackHandleField = CraftItemStack.class.getDeclaredField("handle");
 		craftItemStackHandleField.setAccessible(true);
+
+		// TODO Spigot 1.21.5 remaps "Cow" to "AbstractCow" unless api-version <= "1.21.5". So we
+		// use reflection here to support the new cow variants anyway while still supporting older
+		// api versions.
+		var cowClass = Class.forName("org.bukkit.entity.Cow");
+		cowSetVariantMethod = cowClass.getMethod("setVariant", Cow.Variant.class);
 	}
 
 	@Override
@@ -275,5 +289,79 @@ public final class NMSHandler implements NMSCallProvider {
 			variantValue = Salmon.Variant.MEDIUM; // Default
 		}
 		salmon.setVariant(variantValue);
+	}
+
+	// MC 1.21.5+ TODO Can be removed once we only support Bukkit 1.21.5+
+
+	@Override
+	public void setCowVariant(Cow cow, NamespacedKey variant) {
+		Cow.Variant variantValue = Registry.COW_VARIANT.get(variant);
+		if (variantValue == null) {
+			variantValue = Cow.Variant.TEMPERATE; // Default
+		}
+		assert variantValue != null;
+
+		try {
+			cowSetVariantMethod.invoke(cow, variantValue);
+		} catch (Exception e) {
+			// Unexpected:
+			Log.severe("Failed to set cow variant!", e);
+		}
+	}
+
+	@Override
+	public NamespacedKey cycleCowVariant(NamespacedKey variant, boolean backwards) {
+		Cow.Variant variantValue = Registry.COW_VARIANT.get(variant);
+		if (variantValue == null) {
+			variantValue = Cow.Variant.TEMPERATE; // Default
+		}
+		assert variantValue != null;
+
+		Registry<Cow.Variant> registry = Unsafe.castNonNull(Registry.COW_VARIANT);
+		return RegistryUtils.cycleKeyed(registry, variantValue, backwards).getKeyOrThrow();
+	}
+
+	@Override
+	public void setPigVariant(Pig pig, NamespacedKey variant) {
+		Pig.Variant variantValue = Registry.PIG_VARIANT.get(variant);
+		if (variantValue == null) {
+			variantValue = Pig.Variant.TEMPERATE; // Default
+		}
+		assert variantValue != null;
+		pig.setVariant(variantValue);
+	}
+
+	@Override
+	public NamespacedKey cyclePigVariant(NamespacedKey variant, boolean backwards) {
+		Pig.Variant variantValue = Registry.PIG_VARIANT.get(variant);
+		if (variantValue == null) {
+			variantValue = Pig.Variant.TEMPERATE; // Default
+		}
+		assert variantValue != null;
+
+		Registry<Pig.Variant> registry = Unsafe.castNonNull(Registry.PIG_VARIANT);
+		return RegistryUtils.cycleKeyed(registry, variantValue, backwards).getKeyOrThrow();
+	}
+
+	@Override
+	public void setChickenVariant(Chicken chicken, NamespacedKey variant) {
+		Chicken.Variant variantValue = Registry.CHICKEN_VARIANT.get(variant);
+		if (variantValue == null) {
+			variantValue = Chicken.Variant.TEMPERATE; // Default
+		}
+		assert variantValue != null;
+		chicken.setVariant(variantValue);
+	}
+
+	@Override
+	public NamespacedKey cycleChickenVariant(NamespacedKey variant, boolean backwards) {
+		Chicken.Variant variantValue = Registry.CHICKEN_VARIANT.get(variant);
+		if (variantValue == null) {
+			variantValue = Chicken.Variant.TEMPERATE; // Default
+		}
+		assert variantValue != null;
+
+		Registry<Chicken.Variant> registry = Unsafe.castNonNull(Registry.CHICKEN_VARIANT);
+		return RegistryUtils.cycleKeyed(registry, variantValue, backwards).getKeyOrThrow();
 	}
 }